split family field componenets
